Newer rule for abstractions
Use ``newer'' version of the PTS according to [Pollack 1994], which uses a premise for the validity of the type in the abstractions rules instead of that of the domain and the co-domain (separately).
The change was implemented in both presentation of the PTS, i.e., the finitary and the infinitary one.
It would be better actually to have this change only in the finitary presentation because: first, it would prove that both rules are equivalent, and second; the results in this work wont depart much from the paper, only the interface with the subject reduction repo.